refactor: move tests and fixtures to dedicated files
