+++ crates/p2p/src/client/peer_agnostic.rs
@@ -53,6 +53,11 @@ use tagged::Tagged;
 use tagged_debug_derive::TaggedDebug;
 use tokio::sync::RwLock;
 
+#[cfg(test)]
+mod fixtures;
+#[cfg(test)]
+mod tests;
+
 use crate::client::conv::{CairoDefinition, FromDto, SierraDefinition, TryFromDto};
 use crate::client::peer_aware;
 use crate::sync::protocol;
@@ -1646,560 +1651,6 @@ impl std::fmt::Display for ClassDefinitionsError {
             ClassDefinitionsError::SierraDefinitionError(peer) => {
                 write!(f, "Sierra class definition error from peer {}", peer)
             }
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::collections::VecDeque;
-
-    use fake::{Fake, Faker};
-    use futures::channel::mpsc;
-    use futures::{stream, SinkExt, TryStreamExt};
-    use rstest::rstest;
-    use tagged::Tagged;
-    use tagged_debug_derive::TaggedDebug;
-    use tokio::sync::Mutex;
-
-    use super::*;
-    use crate::client::conv::ToDto;
-
-    #[derive(Clone, PartialEq, TaggedDebug)]
-    struct TestPeer(PeerId);
-
-    fn peer(tag: i32) -> TestPeer {
-        tagged::init();
-        Tagged::<TestPeer>::get(format!("peer {tag}"), || TestPeer(PeerId::random()))
-            .unwrap()
-            .data
-    }
-
-    fn unzip_fixtures<T>(
-        responses: Vec<Result<(TestPeer, Vec<T>), TestPeer>>,
-    ) -> (Vec<PeerId>, Arc<Mutex<VecDeque<Result<Vec<T>, TestPeer>>>>) {
-        let peers = responses
-            .iter()
-            .map(|r| match r {
-                Ok((p, _)) => p.0,
-                Err(p) => p.0,
-            })
-            .collect::<Vec<_>>();
-        let responses = Arc::new(Mutex::new(
-            responses
-                .into_iter()
-                .map(|r| r.map(|(_, responses)| responses))
-                .collect::<VecDeque<_>>(),
-        ));
-        (peers, responses)
-    }
-
-    async fn send_request<T>(
-        responses: Arc<Mutex<VecDeque<Result<Vec<T>, TestPeer>>>>,
-    ) -> anyhow::Result<mpsc::Receiver<T>> {
-        let mut guard = responses.lock().await;
-        match guard.pop_front() {
-            Some(Ok(responses)) => {
-                let (mut tx, rx) = mpsc::channel(responses.len() + 1);
-                for r in responses {
-                    tx.send(r).await.unwrap();
-                }
-                Ok(rx)
-            }
-            Some(Err(_)) => Err(anyhow::anyhow!("peer failed")),
-            None => {
-                panic!("fix your assumed responses")
-            }
-        }
-    }
-
-    mod make_transaction_stream {
-        use TransactionsResponse::Fin;
-
-        use super::*;
-
-        type TestResponse = Result<(TestPeer, Vec<TransactionsResponse>), TestPeer>;
-        #[derive(Clone, Dummy, PartialEq, TaggedDebug)]
-        struct TestTxn {
-            t: TransactionVariant,
-            r: Receipt,
-        }
-
-        #[rstest]
-        #[case::one_peer_1_block(
-            1,
-            // Simulated responses
-            vec![Ok((peer(0), vec![txn_resp(0, 0), txn_resp(1, 1), Fin]))],
-            // Expected number of transactions per block
-            vec![2],
-            // Expected stream
-            vec![Ok((peer(0), vec![txn(0, 0), txn(1, 1)]))]
-        )]
-        #[case::one_peer_2_blocks(
-            // Peer gives responses for all blocks in one go
-            2,
-            vec![Ok((peer(0), vec![txn_resp(2, 0), txn_resp(3, 0), Fin]))],
-            vec![1, 1],
-            vec![
-                Ok((peer(0), vec![txn(2, 0)])), // block 0
-                Ok((peer(0), vec![txn(3, 0)]))  // block 1
-            ]
-        )]
-        #[case::one_peer_2_blocks_in_2_attempts(
-            // Peer gives a response for the second block after a retry
-            2,
-            vec![
-                Ok((peer(0), vec![txn_resp(4, 0), Fin])),
-                Ok((peer(0), vec![txn_resp(5, 0), Fin]))
-            ],
-            vec![1, 1],
-            vec![
-                Ok((peer(0), vec![txn(4, 0)])),
-                Ok((peer(0), vec![txn(5, 0)]))
-            ]
-        )]
-        #[case::two_peers_1_block_per_peer(
-            2,
-            vec![
-                // Errors are ignored
-                Err(peer(1)),
-                Ok((peer(0), vec![txn_resp(6, 0), Fin])),
-                Err(peer(0)),
-                Ok((peer(1), vec![txn_resp(7, 0), Fin]))
-            ],
-            vec![1, 1],
-            vec![
-                Ok((peer(0), vec![txn(6, 0)])),
-                Ok((peer(1), vec![txn(7, 0)]))
-            ]
-        )]
-        #[case::first_peer_premature_eos_with_fin(
-            2,
-            vec![
-                // First peer gives full block 0 and half of block 1
-                Ok((peer(0), vec![txn_resp(8, 0), txn_resp(9, 0), Fin])),
-                Ok((peer(1), vec![txn_resp(9, 0), txn_resp(10, 1), Fin]))
-            ],
-            vec![1, 2],
-            vec![
-                Ok((peer(0), vec![txn(8, 0)])),
-                Ok((peer(1), vec![txn(9, 0), txn(10, 1)]))
-            ]
-        )]
-        #[case::first_peer_full_block_no_fin(
-            2,
-            vec![
-                // First peer gives full block 0 but no fin
-                Ok((peer(0), vec![txn_resp(11, 0)])),
-                Ok((peer(1), vec![txn_resp(12, 0), Fin]))
-            ],
-            vec![1, 1],
-            vec![
-                // We assume this block 0 could be correct
-                Ok((peer(0), vec![txn(11, 0)])), // block 0
-                Ok((peer(1), vec![txn(12, 0)]))  // block 1
-            ]
-        )]
-        // The same as above but the first peer gives half of the second block before closing the
-        // stream
-        #[case::first_peer_half_block_no_fin(
-            2,
-            vec![
-                // First peer gives full block 0 and partial block 1 but no fin
-                Ok((peer(0), vec![txn_resp(13, 0), txn_resp(14, 0)])),
-                Ok((peer(1), vec![txn_resp(14, 0), txn_resp(15, 1), Fin]))
-            ],
-            vec![1, 2],
-            vec![
-                // We assume this block could be correct so we move to the next one
-                Ok((peer(0), vec![txn(13, 0)])),            // block 0
-                Ok((peer(1), vec![txn(14, 0), txn(15, 1)])) // block 1
-            ]
-        )]
-        #[case::count_steam_is_too_short(
-            2,
-            vec![
-                // 2 blocks in responses
-                Ok((peer(0), vec![txn_resp(16, 0), Fin])),
-                Ok((peer(0), vec![txn_resp(17, 0), Fin]))
-            ],
-            vec![1], // but only 1 block provided in the count stream
-            vec![
-                Ok((peer(0), vec![txn(16, 0)])),
-                Err(peer(0)) // the second block is not processed
-            ]
-        )]
-        #[case::too_many_responses(
-            1,
-            vec![Ok((peer(0), vec![txn_resp(18, 0), txn_resp(19, 0), Fin]))],
-            vec![1],
-            vec![Ok((peer(0), vec![txn(18, 0)]))]
-        )]
-        #[test_log::test(tokio::test)]
-        async fn make_transaction_stream(
-            #[case] num_blocks: usize,
-            #[case] responses: Vec<TestResponse>,
-            #[case] num_txns_per_block: Vec<usize>,
-            #[case] expected_stream: Vec<Result<(TestPeer, Vec<TestTxn>), TestPeer>>,
-        ) {
-            let _ = env_logger::builder().is_test(true).try_init();
-            let (peers, responses) = unzip_fixtures(responses);
-            let get_peers = || async { peers.clone() };
-            let send_request =
-                |_: PeerId, _: TransactionsRequest| async { send_request(responses.clone()).await };
-
-            let start = BlockNumber::GENESIS;
-            let stop = start + (num_blocks - 1) as u64;
-
-            let actual = super::make_transaction_stream(
-                start,
-                stop,
-                stream::iter(
-                    num_txns_per_block
-                        .into_iter()
-                        .map(|x| Ok((x, Default::default()))),
-                ),
-                get_peers,
-                send_request,
-            )
-            .map_ok(|x| {
-                (
-                    TestPeer(x.peer),
-                    x.data
-                        .0
-                        .transactions
-                        .into_iter()
-                        .map(|(t, r)| TestTxn { t, r })
-                        .collect(),
-                )
-            })
-            .map_err(|x| TestPeer(x.peer))
-            .collect::<Vec<_>>()
-            .await;
-
-            pretty_assertions_sorted::assert_eq!(actual, expected_stream);
-        }
-
-        fn txn_resp(tag: i32, transaction_index: u64) -> TransactionsResponse {
-            let TestTxn { t, r } = txn(tag, transaction_index);
-            let resp = TransactionsResponse::TransactionWithReceipt(TransactionWithReceipt {
-                receipt: (&t, r).to_dto(),
-                transaction: t.to_dto(),
-            });
-            Tagged::get(format!("txn resp {tag}"), || resp)
-                .unwrap()
-                .data
-        }
-
-        fn txn(tag: i32, transaction_index: u64) -> TestTxn {
-            Tagged::get(format!("txn {tag}"), || {
-                let mut x = Faker.fake::<TestTxn>();
-                x.r.transaction_index = TransactionIndex::new_or_panic(transaction_index);
-                x
-            })
-            .unwrap()
-            .data
-        }
-    }
-
-    mod make_state_diff_stream {
-        use p2p_proto::common::{Address, Hash, VolitionDomain};
-        use p2p_proto::state::{ContractDiff, DeclaredClass};
-        use pathfinder_common::state_update::ContractUpdate;
-        use StateDiffsResponse::Fin;
-
-        use super::*;
-
-        type TestResponse = Result<(TestPeer, Vec<StateDiffsResponse>), TestPeer>;
-
-        #[rstest]
-        #[case::one_peer_1_block(
-            1,
-            vec![Ok((peer(0), vec![contract_diff(0), declared_class(0), Fin]))],
-            vec![len(0)],
-            vec![Ok((peer(0), state_diff(0)))]
-        )]
-        #[case::one_peer_2_blocks(
-            2,
-            vec![Ok((peer(0), vec![contract_diff(0), declared_class(0), contract_diff(1), declared_class(1), Fin]))],
-            vec![len(0), len(1)],
-            vec![Ok((peer(0), state_diff(0))), Ok((peer(0), state_diff(1)))]
-        )]
-        #[case::one_peer_2_blocks_in_2_attempts(
-            // Peer gives a response for the second block after a retry
-            2,
-            vec![
-                Ok((peer(0), vec![contract_diff(0), declared_class(0), Fin])),
-                Ok((peer(0), vec![contract_diff(1), declared_class(1), Fin])),
-            ],
-            vec![len(0), len(1)],
-            vec![
-                Ok((peer(0), state_diff(0))),
-                Ok((peer(0), state_diff(1)))
-            ]
-        )]
-        #[case::two_peers_1_block_per_peer(
-            2,
-            vec![
-                // Errors are ignored
-                Err(peer(1)),
-                Ok((peer(0), vec![contract_diff(0), declared_class(0), Fin])),
-                Err(peer(0)),
-                Ok((peer(1), vec![contract_diff(1), declared_class(1), Fin])),
-            ],
-            vec![len(0), len(1)],
-            vec![
-                Ok((peer(0), state_diff(0))),
-                Ok((peer(1), state_diff(1)))
-            ]
-        )]
-        #[case::first_peer_premature_eos_with_fin(
-            2,
-            vec![
-                // First peer gives full block 0 and half of block 1
-                Ok((peer(0), vec![contract_diff(0), declared_class(0), contract_diff(1), Fin])),
-                Ok((peer(1), vec![contract_diff(1), declared_class(1), Fin]))
-            ],
-            vec![len(0), len(1)],
-            vec![
-                Ok((peer(0), state_diff(0))),
-                Ok((peer(1), state_diff(1)))
-            ]
-        )]
-        #[case::first_peer_full_block_no_fin(
-            2,
-            vec![
-                // First peer gives full block 0 but no fin
-                Ok((peer(0), vec![contract_diff(0), declared_class(0)])),
-                Ok((peer(1), vec![contract_diff(1), declared_class(1), Fin]))
-            ],
-            vec![len(0), len(1)],
-            vec![
-                Ok((peer(0), state_diff(0))),
-                Ok((peer(1), state_diff(1)))
-            ]
-        )]
-        // The same as above but the first peer gives half of the second block before closing the
-        // stream
-        #[case::first_peer_half_block_no_fin(
-            2,
-            vec![
-                // First peer gives full block 0 and partial block 1 but no fin
-                Ok((peer(0), vec![contract_diff(0), declared_class(0), contract_diff(1)])),
-                Ok((peer(1), vec![contract_diff(1), declared_class(1), Fin])),
-            ],
-            vec![len(0), len(1)],
-            vec![
-                Ok((peer(0), state_diff(0))),
-                Ok((peer(1), state_diff(1)))
-            ]
-        )]
-        #[case::count_steam_is_too_short(
-            2,
-            vec![
-                // 2 blocks in responses
-                Ok((peer(0), vec![contract_diff(0), declared_class(0), Fin])),
-                Ok((peer(0), vec![contract_diff(1), declared_class(1), Fin]))
-            ],
-            vec![len(0)], // but only 1 block provided in the count stream
-            vec![
-                Ok((peer(0), state_diff(0))),
-                Err(peer(0)) // the second block is not processed
-            ]
-        )]
-        #[case::too_many_responses_storage(
-            1,
-            vec![Ok((peer(0), vec![contract_diff(0), declared_class(0), surplus_storage(), Fin]))],
-            vec![len(0)],
-            vec![Ok((peer(0), state_diff(0)))]
-        )]
-        #[case::too_many_responses_nonce(
-            1,
-            vec![Ok((peer(0), vec![contract_diff(0), declared_class(0), surplus_nonce(), Fin]))],
-            vec![len(0)],
-            vec![Ok((peer(0), state_diff(0)))]
-        )]
-        #[case::too_many_responses_class(
-            1,
-            vec![Ok((peer(0), vec![contract_diff(0), declared_class(0), surplus_class(), Fin]))],
-            vec![len(0)],
-            vec![Ok((peer(0), state_diff(0)))]
-        )]
-        #[case::too_many_responses_declaration(
-            1,
-            vec![Ok((peer(0), vec![contract_diff(0), declared_class(0), declared_class(1), Fin]))],
-            vec![len(0)],
-            vec![Ok((peer(0), state_diff(0)))]
-        )]
-        #[test_log::test(tokio::test)]
-        async fn make_state_diff_stream(
-            #[case] num_blocks: usize,
-            #[case] responses: Vec<TestResponse>,
-            #[case] state_diff_len_per_block: Vec<usize>,
-            #[case] expected_stream: Vec<Result<(TestPeer, UnverifiedStateUpdateData), TestPeer>>,
-        ) {
-            let _ = env_logger::builder().is_test(true).try_init();
-            let (peers, responses) = unzip_fixtures(responses);
-            let get_peers = || async { peers.clone() };
-            let send_request =
-                |_: PeerId, _: StateDiffsRequest| async { send_request(responses.clone()).await };
-
-            let start = BlockNumber::GENESIS;
-            let stop = start + (num_blocks - 1) as u64;
-
-            let actual = super::make_state_diff_stream(
-                start,
-                stop,
-                stream::iter(
-                    state_diff_len_per_block
-                        .into_iter()
-                        .map(|x| Ok((x, Default::default()))),
-                ),
-                get_peers,
-                send_request,
-            )
-            .map_ok(|x| (TestPeer(x.peer), x.data))
-            .map_err(|x| TestPeer(x.peer))
-            .collect::<Vec<_>>()
-            .await;
-
-            // Add block numbers
-            let expected = expected_stream
-                .into_iter()
-                .enumerate()
-                .map(|(i, x)| x.map(|(p, su)| (p, (su, BlockNumber::new_or_panic(i as u64)))))
-                .collect::<Vec<_>>();
-
-            pretty_assertions_sorted::assert_eq!(actual, expected);
-        }
-
-        fn contract_diff(tag: i32) -> StateDiffsResponse {
-            let sd = state_diff(tag).state_diff;
-            let (a, u) = sd
-                .contract_updates
-                .into_iter()
-                .chain(sd.system_contract_updates.into_iter().map(|(a, u)| {
-                    (
-                        a,
-                        ContractUpdate {
-                            storage: u.storage,
-                            ..Default::default()
-                        },
-                    )
-                }))
-                .next()
-                .unwrap();
-            StateDiffsResponse::ContractDiff(
-                Tagged::get(format!("contract diff response {tag}"), || ContractDiff {
-                    address: Address(a.0),
-                    nonce: u.nonce.map(|x| x.0),
-                    class_hash: u.class.map(|x| Hash(x.class_hash().0)),
-                    values: u
-                        .storage
-                        .into_iter()
-                        .map(|(k, v)| ContractStoredValue {
-                            key: k.0,
-                            value: v.0,
-                        })
-                        .collect(),
-                    domain: VolitionDomain::L1,
-                })
-                .unwrap()
-                .data,
-            )
-        }
-
-        fn declared_class(tag: i32) -> StateDiffsResponse {
-            let sd = state_diff(tag).state_diff;
-            let (class_hash, compiled_class_hash) = sd
-                .declared_sierra_classes
-                .into_iter()
-                .map(|(s, c)| (Hash(s.0), Some(Hash(c.0))))
-                .chain(
-                    sd.declared_cairo_classes
-                        .into_iter()
-                        .map(|c| (Hash(c.0), None)),
-                )
-                .next()
-                .unwrap();
-
-            StateDiffsResponse::DeclaredClass(
-                Tagged::get(format!("declared class {tag}"), || DeclaredClass {
-                    class_hash,
-                    compiled_class_hash,
-                })
-                .unwrap()
-                .data,
-            )
-        }
-
-        fn state_diff(tag: i32) -> UnverifiedStateUpdateData {
-            let (declared_cairo_classes, declared_sierra_classes) =
-                match Faker.fake::<Option<CasmHash>>() {
-                    Some(x) => ([].into(), [(SierraHash(Faker.fake()), x)].into()),
-                    None => ([ClassHash(Faker.fake())].into(), [].into()),
-                };
-            let (contract_updates, system_contract_updates) = if Faker.fake() {
-                (
-                    [(
-                        Faker.fake(),
-                        ContractUpdate {
-                            storage: Faker.fake(),
-                            class: Some(ContractClassUpdate::Deploy(Faker.fake())),
-                            nonce: Faker.fake(),
-                        },
-                    )]
-                    .into(),
-                    [].into(),
-                )
-            } else {
-                ([].into(), [(ContractAddress::ONE, Faker.fake())].into())
-            };
-            Tagged::get(format!("state diff {tag}"), || UnverifiedStateUpdateData {
-                expected_commitment: Default::default(),
-                state_diff: StateUpdateData {
-                    contract_updates,
-                    system_contract_updates,
-                    declared_cairo_classes,
-                    declared_sierra_classes,
-                },
-            })
-            .unwrap()
-            .data
-        }
-
-        fn len(tag: i32) -> usize {
-            state_diff(tag).state_diff.state_diff_length()
-        }
-
-        fn surplus_storage() -> StateDiffsResponse {
-            StateDiffsResponse::ContractDiff(ContractDiff {
-                address: Faker.fake(),
-                nonce: None,
-                class_hash: None,
-                values: vec![Faker.fake()], // Must not be empty
-                domain: Faker.fake(),
-            })
-        }
-
-        fn surplus_nonce() -> StateDiffsResponse {
-            StateDiffsResponse::ContractDiff(ContractDiff {
-                address: Faker.fake(),
-                nonce: Some(Faker.fake()),
-                class_hash: None,
-                values: vec![],
-                domain: Faker.fake(),
-            })
-        }
-
-        fn surplus_class() -> StateDiffsResponse {
-            StateDiffsResponse::ContractDiff(ContractDiff {
-                address: Faker.fake(),
-                nonce: None,
-                class_hash: Some(Faker.fake()),
-                values: vec![],
-                domain: Faker.fake(),
-            })
         }
     }
 }
